@@ -1,13 +1,14 @@
 ﻿using Microsoft.AspNetCore.SignalR;
+using System.Threading.Tasks;
 
 namespace ChatSample.Hubs
 {
     public class ChatHub : Hub
     {
-        public void Send(string name, string message)
+        public async Task Send(string name, string message)
         {
             // Call the broadcastMessage method to update clients.
-            Clients.All.SendAsync("broadcastMessage", name, message);
+            await Clients.All.SendAsync("broadcastMessage", name, message);
         }
     }
 }
